+++ .env
@@ -1,2 +1,2 @@
 VITE_APP_TITLE = Togeeat
-VITE_APP_API_URL = http://localhost:3000/api/v1
+VITE_APP_API_URL = https://togeeat-be-production.up.railway.app/api/v1
