+++ src/components/bussiness/ScheduleMatchingItem.tsx
@@ -1,6 +1,6 @@
-import StonksImage from "@/assets/images/stonks.jpg";
 import UserProfilePopup from "@/components/bussiness/UserProfilePopup";
-import { Avatar, Image, Space, Tooltip } from "antd";
+import { GoogleMap, MarkerF } from "@react-google-maps/api";
+import { Avatar, Space, Tooltip } from "antd";
 import dayjs from "dayjs";
 import { useState } from "react";
 import { useTranslation } from "react-i18next";
@@ -17,7 +17,21 @@ function ScheduleMatchingItem(props: Props) {
     <div className="flex flex-row justify-between p-4 bg-white rounded-lg">
       <div className="flex flex-col">
         <div className="flex flex-row gap-6">
-          <Image preview={false} src={StonksImage} width={240} />
+          <GoogleMap
+            mapContainerStyle={{ width: 240, height: 160 }}
+            center={{ lat: props.lat, lng: props.long }}
+            zoom={15}
+          >
+            <MarkerF position={{ lat: props.lat, lng: props.long }} />
+          </GoogleMap>
+          {/* <Image preview={false} src={StonksImage} width={240} /> */}
+          {/* <iframe
+            src="https://www.google.com/maps/embed?pb=!1m18!1m12!1m3!1d2851.5163583264534!2d105.8445778242964!3d21.002423480742486!2m3!1f0!2f0!3f0!3m2!1i1024!2i768!4f13.1!3m3!1m2!1s0x3135adb4871c8657%3A0xf8ed43d02a44ae39!2sCLB%20Bi-A%20B%C3%A1ch%20Khoa!5e0!3m2!1sen!2s!4v1687273558026!5m2!1sen!2s"
+            width="240"
+            height="160"
+            style={{ border: 0 }}
+            loading="lazy"
+          ></iframe> */}
           <div className="flex flex-col items-start gap-4">
             <p className="text-2xl font-semibold">{props.address}</p>
             <div>{t("matching.text.participant")}:</div>
